+++ src/U_stats/statistics/U_statistics.py
@@ -1,33 +1,29 @@
-from ..tensor_contraction.path import TensorExpression, NestedHashableList
+from ..tensor_contraction.path import TensorExpression
 from ..tensor_contraction.calculator import TensorContractionCalculator
 from typing import List, Hashable, Generator, Tuple, Dict, Set
 from functools import cached_property
-import warnings
 from .U2V import get_all_partitions, partition_weight, get_all_partitions_nonconnected
-from ..utils import standardize_indexes
+from .._utils import Expression, standardize_indices
 import numpy as np
 import itertools
 
-try:
-    import torch
-except ImportError:
-    warnings.warn(
-        "Torch is not installed. Tensor contraction will be performed using NumPy."
-    )
-__all__ = ["UStatsCalculator", "U_stats"]
+__all__ = [
+    "UStatsCalculator",
+    "U_stats_loop",
+]
 
 
 class UExpression(TensorExpression):
 
-    def __init__(self, mode: NestedHashableList):
-        super().__init__(mode)
+    def __init__(self, expression: Expression):
+        super().__init__(expression)
 
     @property
     def order(self) -> int:
         return len(self.indices)
 
-    def submode(self, partition: List[set]) -> Tuple[float, "UExpression"]:
-        new_mode = self.copy()
+    def subexpression(self, partition: List[set]) -> Tuple[float, "UExpression"]:
+        new_expression = self.copy()
         mapping = {}
         weight = partition_weight(partition)
 
@@ -36,14 +32,14 @@ class UExpression(TensorExpression):
             for element in s:
                 if element != representative:
                     mapping[element] = representative
-        for pos in new_mode._pair_dict:
-            pair = new_mode[pos]
+        for pos in new_expression._pair_dict:
+            pair = new_expression[pos]
             new_pair = [mapping.get(index, index) for index in pair]
-            new_mode._pair_dict[pos] = new_pair
+            new_expression._pair_dict[pos] = new_pair
 
         for orig, rep in mapping.items():
-            new_mode._index_table.merge(orig, rep)
-        return weight, new_mode
+            new_expression._index_table.merge(orig, rep)
+        return weight, new_expression
 
     @cached_property
     def _adj_list(self) -> Dict[Hashable, Set[Hashable]]:
@@ -56,23 +52,25 @@ class UExpression(TensorExpression):
             adj_list[index].discard(index)
         return adj_list
 
-    def submodes(self) -> Generator[Tuple[float, "UExpression"], None, None]:
+    def subexpressions(self) -> Generator[Tuple[float, "UExpression"], None, None]:
         partitions = get_all_partitions(self._index_table.indices)
         for partition in partitions:
-            yield self.submode(partition)
+            yield self.subexpression(partition)
 
-    def non_diag_submodes(self) -> Generator[Tuple[float, "UExpression"], None, None]:
+    def non_diag_subexpressions(
+        self,
+    ) -> Generator[Tuple[float, "UExpression"], None, None]:
         partitions = get_all_partitions_nonconnected(self._adj_list)
         for partition in partitions:
-            weight, submode = self.submode(partition)
-            yield weight, submode
+            weight, subexpression = self.subexpression(partition)
+            yield weight, subexpression
 
 
 class UStatsCalculator(TensorContractionCalculator):
     """A class for calculating the U statistics of a list of kernel
-    matrices(tensors) with particular mode."""
+    matrices(tensors) with particular expression."""
 
-    def __init__(self, mode: NestedHashableList, summor: str = "numpy"):
+    def __init__(self, expression: Expression, summor: str = "numpy"):
         """Initialize UStatsCalculator with specified tensor contraction
         backend.
 
@@ -80,15 +78,15 @@ class UStatsCalculator(TensorContractionCalculator):
             summor: str, either "numpy" or "torch"
         """
         TensorContractionCalculator.__init__(self, summor)
-        self.mode = UExpression(mode)
-        self.order = self.mode.order
-        self.shape = self.mode.shape
+        self.expression = UExpression(expression)
+        self.order = self.expression.order
+        self.shape = self.expression.shape
 
     def calculate(
         self, tensors: List[np.ndarray], average=True, path_method="greedy"
     ) -> float:
         """Calculate the U statistics of a list of kernel matrices(tensors)
-        with particular mode.
+        with particular expression.
 
         Args:
             tensors: List[np.ndarray], a list of kernel matrices
@@ -99,9 +97,9 @@ class UStatsCalculator(TensorContractionCalculator):
         tensors = self._initalize_tensor_dict(tensors, self.shape)
         self._validate_inputs(tensors, self.shape)
         result = 0
-        submodes = self.mode.submodes()
-        for weight, submode in submodes:
-            path, cost = submode.path(path_method)
+        subexpressions = self.expression.subexpressions()
+        for weight, subexpression in subexpressions:
+            path, cost = subexpression.path(path_method)
             result += weight * TensorContractionCalculator._tensor_contract(
                 self, tensors.copy(), path
             )
@@ -114,7 +112,7 @@ class UStatsCalculator(TensorContractionCalculator):
         self, tensors: List[np.ndarray], average=True, path_method="greedy"
     ) -> float:
         """Calculate the U statistics of a list of kernel matrices(tensors)
-        with particular mode.
+        with particular expression.
 
         Args:
             tensors: List[np.ndarray], a list of kernel matrices
@@ -131,9 +129,9 @@ class UStatsCalculator(TensorContractionCalculator):
             tensors = self._torch_mask_tensors(tensors, n_samples, self.device)
 
         result = 0
-        submodes = self.mode.non_diag_submodes()
-        for weight, submode in submodes:
-            path, cost = submode.path(path_method)
+        subexpressions = self.expression.non_diag_subexpressions()
+        for weight, subexpression in subexpressions:
+            path, cost = subexpression.path(path_method)
             result += weight * TensorContractionCalculator._tensor_contract(
                 self, tensors.copy(), computing_path=path
             )
@@ -169,67 +167,12 @@ class UStatsCalculator(TensorContractionCalculator):
         mask = np.broadcast_to(mask, (dim,) * ndim)
         return mask
 
-    @staticmethod
-    def _torch_mask_tensors(
-        tensors: Dict[int, torch.Tensor], sample_size: int, device: torch.device
-    ) -> Dict[int, torch.Tensor]:
-        shapes = {index: tensor.dim() for index, tensor in tensors.items()}
-        for index, ndim in shapes.items():
-            if ndim > 1:
-                mask_total = torch.zeros(
-                    (sample_size,) * ndim, dtype=torch.bool, device=device
-                )
-                for i, j in itertools.combinations(range(ndim), 2):
-                    mask = UStatsCalculator._torch_mask_tensor(
-                        ndim, sample_size, i, j, device
-                    )
-                    mask_total |= mask
-                mask_total = ~mask_total
-                tensors[index] = tensors[index] * mask_total
-        return tensors
 
-    @staticmethod
-    def _torch_mask_tensor(
-        ndim: int, dim: int, index1: int, index2: int, device: torch.device
-    ) -> torch.Tensor:
-        shape1 = [1] * ndim
-        shape1[index1] = dim
-        shape2 = [1] * ndim
-        shape2[index2] = dim
-
-        idx1 = torch.arange(dim, device=device).reshape(shape1)
-        idx2 = torch.arange(dim, device=device).reshape(shape2)
-        mask = idx1 == idx2
-        mask = mask.broadcast_to((dim,) * ndim)
-        return mask
-
-
-def U_stats(
-    tensors: List[np.ndarray],
-    mode: NestedHashableList,
-    average: bool = True,
-    summor: str = "numpy",
-) -> float:
-    """Calculate the U statistics of a list of kernel matrices(tensors) with
-    particular mode.
-
-    Args:
-        tensors: List[np.ndarray], a list of kernel matrices
-        mode: List[Union[List[Hashable], str]], the mode of the U statistics
-        average: bool, whether to average the U statistics
-        summor: str, either "numpy" or "torch"
-
-    Returns:
-        float, the U statistics of the kernel matrices
-    """
-    return UStatsCalculator(mode, summor=summor).calculate(tensors, average)
-
-
-def U_stats_loop(tensors: List[np.ndarray], mode: List[List[int]]) -> float:
+def U_stats_loop(tensors: List[np.ndarray], expression: List[List[int]]) -> float:
     nt = len(tensors)
     ns = tensors[0].shape[0]
-    mode = standardize_indexes(mode)
-    order = len(set(itertools.chain(*mode)))
+    expression = standardize_indices(expression)
+    order = len(set(itertools.chain(*expression)))
 
     num_perms = np.prod(np.arange(ns, ns - order, -1))
     total_sum = 0.0
@@ -237,7 +180,7 @@ def U_stats_loop(tensors: List[np.ndarray], mode: List[List[int]]) -> float:
     for indices in itertools.permutations(range(ns), order):
         product = 1.0
         for i in range(nt):
-            current_indices = tuple(indices[j] for j in mode[i])
+            current_indices = tuple(indices[j] for j in expression[i])
             product *= tensors[i][current_indices]
         total_sum += product