+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.2.1 (2025-06-07)
+
+### Refactor
+
+- **__init__.py**: add interface for U_stats_loop
+
+## 0.2.0 (2025-06-07)
+
+### Feat
+
+- **all**: add function analyzing expression; add interface in __init__.py; complete meta information in __init__.py
+
+### Fix
+
+- **V_stats**: fix a bug
+- fix chinese checke script name
+- **no_chinese_check.py**: fix script path
+- **path.py**: fix typos and bugs
+
 ## 0.1.0 (2025-05-30)
 
 ### Feat